snap: introduce snap_init_device/snap_teardown_device
These API functions will be used by libsnap_nvme or alike to initialize
and destroy the device resources. E.g in Bluefield-1, one should create
all the device emulation resources on top of the emulated PCI function
(using the vhca tunnel). Implement the initial resource
creation/destruction and use it in the NVMe library. Also create a unit
test to exercise this code.

+++ src/mlx5_snap.h
@@ -49,7 +49,7 @@ struct mlx5_snap_pci {
 
 struct mlx5_snap_devx_obj {
 	struct mlx5dv_devx_obj		*obj;
-	u8				obj_id;
+	u16				obj_id;
 };
 
 struct mlx5_snap_device {
